Initial Commit
Add AppUser package with classes for AppUser where the AppUser object is
defined.  The AppUserRepository which is an interface for CRUD operations
and access to the database - it is the DAO.  The AppUserService is the
service layer which is responsible for fetching and retreiving data
from the repository. Business logic will reside here. ApUserRole class
defines what a users rolls can be.